requirements.txt
pytest
requests
allure-pytest
selenium